some reorg for making more sense
