+++ stats/stats/src/charts/counters/last_new_verified_contracts.rs
@@ -3,7 +3,7 @@ use crate::{
         create_chart,
         db_interaction::{
             types::DateValue,
-            updater::{last_point, ChartDependentUpdater, ChartUpdater},
+            chart_updaters::{last_point, ChartDependentUpdater, ChartUpdater},
         },
     },
     lines::NewVerifiedContracts,
