+++ stats/stats/src/charts/db_interaction/chart_updaters/batch.rs
@@ -6,12 +6,12 @@
 //! `force_full` or initial updates) is too expensive.
 
 use super::{
-    common_operations::read::{get_last_row, get_min_block_blockscout, get_min_date_blockscout},
+    common_operations::{get_last_row, get_min_block_blockscout, get_min_date_blockscout},
     ChartUpdater,
 };
 use crate::{
     charts::{
-        db_interaction::{insert::insert_data_many, types::DateValue},
+        db_interaction::{types::DateValue, write::insert_data_many},
         find_chart,
     },
     metrics, UpdateError,
