--- stats/stats/src/charts/db_interaction/updater/common_operations/mod.rs
@@ -1,5 +1,0 @@
-//! Collection of common operations to perform while updating.
-//! Can be useful for any chart regardless of their type.
-
-pub mod read;
-pub mod update;
