+++ stats/stats/src/charts/db_interaction/write.rs
