+++ stats/stats/src/charts/lines/accounts_growth.rs
@@ -4,7 +4,7 @@ use crate::{
         cache::Cache,
         db_interaction::{
             types::{DateValue, DateValueInt},
-            updater::{ChartFullUpdater, ChartUpdater},
+            chart_updaters::{ChartFullUpdater, ChartUpdater},
         },
     },
     MissingDatePolicy, UpdateError,
