+++ stats/stats/src/charts/lines/active_accounts.rs
@@ -1,7 +1,7 @@
 use crate::{
     charts::db_interaction::{
+        chart_updaters::{ChartPartialUpdater, ChartUpdater},
         types::DateValue,
-        updater::{ChartPartialUpdater, ChartUpdater},
     },
     UpdateError,
 };
