+++ stats/stats/src/charts/lines/average_block_size.rs
@@ -1,7 +1,7 @@
 use crate::{
     charts::db_interaction::{
         types::DateValue,
-        updater::{ChartPartialUpdater, ChartUpdater},
+        chart_updaters::{ChartPartialUpdater, ChartUpdater},
     },
     UpdateError,
 };
