+++ stats/stats/src/charts/lines/native_coin_holders_growth.rs
@@ -2,12 +2,12 @@ use crate::{
     charts::{
         create_chart,
         db_interaction::{
-            insert::insert_data_many,
-            types::DateValue,
-            updater::{
-                common_operations::read::{get_last_row, get_min_block_blockscout},
+            chart_updaters::{
+                common_operations::{get_last_row, get_min_block_blockscout},
                 ChartUpdater,
             },
+            types::DateValue,
+            write::insert_data_many,
         },
         find_chart,
     },
