+++ stats/stats/src/charts/lines/txns_growth.rs
@@ -4,8 +4,8 @@ use crate::{
         chart::Chart,
         create_chart,
         db_interaction::{
+            chart_updaters::{parse_and_cumsum, ChartDependentUpdater, ChartUpdater},
             types::DateValue,
-            updater::{parse_and_cumsum, ChartDependentUpdater, ChartUpdater},
         },
     },
     MissingDatePolicy, UpdateError,
