+++ stats/stats/src/charts/mod.rs
@@ -5,4 +5,4 @@ pub mod db_interaction;
 pub mod lines;
 mod mutex;
 pub use chart::{create_chart, find_chart, Chart, MissingDatePolicy, UpdateError};
-pub use db_interaction::updater::ChartUpdater;
+pub use db_interaction::chart_updaters::ChartUpdater;
